+++ src/core/DDVector2.cpp
@@ -2,7 +2,7 @@
 
 DDVector2::DDVector2() : x{0}, y{0} {}
 
-DDVector2::DDVector2(int x, int y) : x{x}, y{y}
+DDVector2::DDVector2(float x, float y) : x{x}, y{y}
 {
     this->x = x;
     this->y = y;
